# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/guides": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all guides with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guides"
                ],
                "summary": "List All Guides",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guide Title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Guide Content",
                        "name": "content",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Guides",
                        "schema": {
                            "$ref": "#/definitions/genprotos.ListAllGuidesResponse"
                        }
                    },
                    "401": {
                        "description": "Error while listing guides",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing guide",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guides"
                ],
                "summary": "Update Guide",
                "parameters": [
                    {
                        "description": "Guide Data",
                        "name": "guide",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genprotos.UpdateGuideRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guide Updated Successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.Void"
                        }
                    },
                    "401": {
                        "description": "Error while updating guide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new guide",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guides"
                ],
                "summary": "Create Guide",
                "parameters": [
                    {
                        "description": "Guide Data",
                        "name": "guide",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genprotos.CreateGuideRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guide Created Successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.Void"
                        }
                    },
                    "401": {
                        "description": "Error while creating guide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/guides/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search for guides by title or content",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guides"
                ],
                "summary": "Search Guides",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guide Title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Guide Content",
                        "name": "content",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Found Guides",
                        "schema": {
                            "$ref": "#/definitions/genprotos.ListAllGuidesResponse"
                        }
                    },
                    "401": {
                        "description": "Error while searching guides",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/guides/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a guide by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guides"
                ],
                "summary": "Get Guide",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guide ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guide Data",
                        "schema": {
                            "$ref": "#/definitions/genprotos.GuideResponse"
                        }
                    },
                    "401": {
                        "description": "Error while fetching guide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a guide by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guides"
                ],
                "summary": "Delete Guide",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guide ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guide Deleted Successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.Void"
                        }
                    },
                    "401": {
                        "description": "Error while deleting guide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "genprotos.CreateGuideRequest": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "genprotos.Guide": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "genprotos.GuideResponse": {
            "type": "object",
            "properties": {
                "guide": {
                    "$ref": "#/definitions/genprotos.Guide"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "genprotos.ListAllGuidesResponse": {
            "type": "object",
            "properties": {
                "guides": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genprotos.Guide"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "genprotos.UpdateGuideRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "genprotos.Void": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}